+++ apps/MusicLabelApp/backend/config/secrets/prod/prod.list.php
@@ -1,0 +1,7 @@
+<?php
+
+return array (
+  'CONTENTS' => NULL,
+  'IDENTIFIED_BY' => NULL,
+  'PASSPHRASE' => NULL,
+);