+++ src/MusicLabel/Catalog/Application/Album/Get/Collection/AlbumsSearcher.php
@@ -8,7 +8,7 @@ use Masfernandez\MusicLabel\Catalog\Application\Album\Get\AlbumsResponse;
 use Masfernandez\MusicLabel\Catalog\Application\Album\Get\Criteria;
 use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\AlbumRepository;
 use Masfernandez\Shared\Application\Service\ApplicationServiceInterface;
-use Masfernandez\Shared\Domain\Bus\Request\RequestInterface;
+use Masfernandez\Shared\Domain\Bus\Request\Request;
 
 final class AlbumsSearcher implements ApplicationServiceInterface
 {
@@ -16,7 +16,7 @@ final class AlbumsSearcher implements ApplicationServiceInterface
     {
     }
 
-    public function execute(GetAlbumsQuery | RequestInterface $request): AlbumsResponse
+    public function execute(GetAlbumsQuery|Request $request): AlbumsResponse
     {
         $criteria = new Criteria(
             $request->getFilters(), //@todo not implemented