+++ src/MusicLabel/Catalog/Application/Album/Post/PostAlbumCommandHandler.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Masfernandez\MusicLabel\Catalog\Application\Album\Post;
 
 use Masfernandez\Shared\Application\Service\ApplicationServiceInterface;
-use Masfernandez\Shared\Domain\Bus\Command\CommandHandlerInterface;
+use Masfernandez\Shared\Domain\Bus\Command\CommandHandler;
 
-final class PostAlbumCommandHandler implements CommandHandlerInterface
+final class PostAlbumCommandHandler implements CommandHandler
 {
     public function __construct(private ApplicationServiceInterface $albumCreator)
     {