+++ src/MusicLabel/Catalog/Domain/Model/Album/AlbumAlreadyExists.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\MusicLabel\Catalog\Domain\Model\Album;
+
+use Exception;
+
+final class AlbumAlreadyExists extends Exception
+{
+    // @todo exception handler and map all domain exceptions
+}