+++ src/Shared/Application/Service/ApplicationServiceInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Masfernandez\Shared\Application\Service;
 
-use Masfernandez\Shared\Domain\Bus\Request\RequestInterface;
+use Masfernandez\Shared\Domain\Bus\Request\Request;
 
 interface ApplicationServiceInterface
 {
-    public function execute(RequestInterface $request);
+    public function execute(Request $request);
 }