+++ src/Shared/Domain/Model/DomainException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\Shared\Domain\Model;
+
+use Exception;
+
+class DomainException extends Exception
+{
+}