+++ src/Shared/Infrastructure/Bus/Event/InMemory/InMemoryEventPublisher.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace Masfernandez\Shared\Infrastructure\Bus\Event\InMemory;
 
-use Masfernandez\Shared\Domain\Bus\Event\DomainEventAbstract;
-use Masfernandez\Shared\Domain\Bus\Event\EventPublisherInterface;
+use Masfernandez\Shared\Domain\Bus\Event\DomainEvent;
+use Masfernandez\Shared\Domain\Bus\Event\EventPublisher;
 use Symfony\Component\Messenger\MessageBusInterface;
 
-final class InMemoryEventPublisher implements EventPublisherInterface
+final class InMemoryEventPublisher implements EventPublisher
 {
     public function __construct(protected MessageBusInterface $eventBus)
     {
     }
 
-    public function publish(DomainEventAbstract ...$events): void
+    public function publish(DomainEvent ...$events): void
     {
         foreach ($events as $event) {
             $this->eventBus->dispatch($event);