+++ tests/MusicLabel/Catalog/Application/Album/Get/Single/AlbumSearcherTest.php
@@ -9,7 +9,7 @@ namespace Masfernandez\Tests\MusicLabel\Catalog\Application\Album\Get\Single;
 use Masfernandez\MusicLabel\Catalog\Application\Album\Get\AlbumResponse;
 use Masfernandez\MusicLabel\Catalog\Application\Album\Get\Single\AlbumSearcher;
 use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\AlbumRepository;
-use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\InMemoryRepository;
+use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\CacheInMemory;
 use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumMother;
 use Mockery;
 use PHPUnit\Framework\TestCase;
@@ -29,7 +29,7 @@ class AlbumSearcherTest extends TestCase
         $albumRepository = Mockery::mock(AlbumRepository::class);
         $albumRepository->allows()->getById($query->id())->andReturns($album);
 
-        $inMemoryRepository = Mockery::mock(InMemoryRepository::class);
+        $inMemoryRepository = Mockery::mock(CacheInMemory::class);
         $inMemoryRepository->expects()
             ->get($query->id()->toString())
             ->andReturns(json_encode($album->toArray(), JSON_THROW_ON_ERROR));