+++ tests/MusicLabel/Catalog/Application/Album/Patch/AlbumPatcherTest.php
@@ -8,7 +8,7 @@ namespace Masfernandez\Tests\MusicLabel\Catalog\Application\Album\Patch;
 
 use Masfernandez\MusicLabel\Catalog\Application\Album\Patch\AlbumPatcher;
 use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\AlbumRepository;
-use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\InMemoryRepository;
+use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\CacheInMemory;
 use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumMother;
 use Mockery;
 use PHPUnit\Framework\TestCase;
@@ -28,7 +28,7 @@ class AlbumPatcherTest extends TestCase
         $albumRepository->allows()->getById($command->getId())->andReturns($album);
         $albumRepository->expects()->patch($album);
 
-        $inMemoryRepository = Mockery::mock(InMemoryRepository::class);
+        $inMemoryRepository = Mockery::mock(CacheInMemory::class);
         $inMemoryRepository->allows()->set(
             $command->getId()->toString(),
             json_encode($album->toArray(), JSON_THROW_ON_ERROR)