+++ CHANGELOG.md
@@ -14,6 +14,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.3.0] - 2021-04-01
+### Added
+- JWT authorization for endpoints under /albums wit PUT verb
+- Domain exceptions wrapper to HTTP codes 
+### Changed
+- Acceptation and Unit tests for a new feature added
+- Documentation
+- Delete some *Exception *Interfaces suffixes from class names
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
+
 ## [0.2.3] - 2021-01-28
 ### Added
 - DumpEnvCommand to generate .env.local.php properly
@@ -114,6 +129,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [Unreleased]: https://github.com/masfernandez/symfony-ddd-hexarch-cqrs/compare/master...develop
+[0.3.0]: https://github.com/masfernandez/symfony-ddd-hexarch-cqrs/compare/v0.2.3...v0.3.0
 [0.2.3]: https://github.com/masfernandez/symfony-ddd-hexarch-cqrs/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/masfernandez/symfony-ddd-hexarch-cqrs/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/masfernandez/symfony-ddd-hexarch-cqrs/compare/v0.2.0...v0.2.1